refactor: rebuild db service
